Add files via upload
The Heart Disease Prediction project is a machine learning application designed to predict the likelihood of heart disease in individuals based on various health metrics and lifestyle factors. By leveraging data analysis and predictive modeling techniques, this project aims to provide insights that can help in early diagnosis and prevention of heart-related issues.